requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
bcrypt==3.2.0
redis==3.5.3
pyOpenSSL==20.0.1
requests==2.31.0
python-dotenv==0.19.0
flask-limiter
